% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kneedle.R
\name{kneedle}
\alias{kneedle}
\title{kneedle
This function allows you to find the knee of a graph}
\usage{
kneedle(x, y, decreasing, concave, sensitivity = 1)
}
\arguments{
\item{x}{A vector of x coordinates.}

\item{y}{A vector of y coordinates.}

\item{decreasing}{Is the function increasing? Algorithm will take a guess at direction of data if not specified.}

\item{concave}{Is this concave or convex? Algorithm will take a guess at concavity if not specified.}

\item{sensitivity}{How sensitive should knee detection be? Defaults to 1. Can be a float/double.}
}
\value{
This function returns a x,y coordinate pair that corresponds to the knee point
}
\description{
kneedle
This function allows you to find the knee of a graph
}
\examples{
x <- kneedle(c(1,2,3,4,5), c(0,1,2,40,60))
}
\keyword{knee}
