% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kneedle.R
\name{kneedle}
\alias{kneedle}
\title{Kneedle}
\usage{
kneedle(x, y, decreasing, sensitivity, concave)
}
\arguments{
\item{y}{}
}
\description{
This function allows you to find the knee of a graph
}
\examples{
kneedle()
}
\keyword{knee}
