% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kneedle.R
\name{kneedle}
\alias{kneedle}
\title{kneedle
This function allows you to find the knee of a graph}
\usage{
kneedle(x, y, decreasing = FALSE, sensitivity = 1, concave = TRUE)
}
\arguments{
\item{x}{A vector of x coordinates.}

\item{y}{A vector of y coordinates.}

\item{decreasing}{Is the function increasing? Defaults to FALSE.}

\item{sensitivity}{How sensitive should knee detection be? Defaults to 1. Can be a float/double.}

\item{concave}{Is this concave or convex? Defaults to TRUE.}
}
\value{
This function returns a x,y coordinate pair that corresponds to the knee point
kneedle()
}
\description{
kneedle
This function allows you to find the knee of a graph
}
\keyword{knee}
